requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.13
frozenlist==1.7.0
h11==0.16.0
icalendar==6.3.2
idna==3.10
multidict==6.5.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
redis==6.2.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
yarl==1.20.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
